refine the codes
