+++ research/query_service/ir/graph_proxy/src/apis/graph/element/vertex.rs
@@ -24,8 +24,7 @@ use ir_common::generated::results as result_pb;
 use ir_common::NameOrId;
 use pegasus_common::codec::{Decode, Encode, ReadExt, WriteExt};
 
-use crate::api::graph::element::{DefaultDetails, DynDetails, Element, GraphElement};
-use crate::api::graph::{read_id, write_id, ID};
+use crate::apis::{read_id, write_id, DefaultDetails, DynDetails, Element, GraphElement, ID};
 use crate::utils::expr::eval::Context;
 
 #[derive(Clone, Debug)]
