+++ research/query_service/ir/graph_proxy/src/read_adapters/exp_store/graph_partition.rs
@@ -13,8 +13,7 @@
 //! See the License for the specific language governing permissions and
 //! limitations under the License.
 
-use crate::api::graph::ID;
-use crate::api::partitioner::Partitioner;
+use crate::apis::{Partitioner, ID};
 use crate::errors::GraphProxyResult;
 
 /// A simple partition utility that one server contains a single graph partition
