+++ research/query_service/ir/graph_proxy/src/read_adapters/gs_store/graph_partition.rs
@@ -19,8 +19,7 @@ use std::sync::{Arc, RwLock};
 use maxgraph_store::api::graph_partition::GraphPartitionManager;
 use maxgraph_store::api::{PartitionId, VertexId};
 
-use crate::api::graph::ID;
-use crate::api::partitioner::Partitioner;
+use crate::apis::{Partitioner, ID};
 use crate::errors::{GraphProxyError, GraphProxyResult};
 
 /// A partition utility that one server contains multiple graph partitions for Groot Store
