+++ research/query_service/ir/integrated/tests/auxilia_test.rs
@@ -21,7 +21,7 @@ mod common;
 mod test {
     use std::sync::Arc;
 
-    use graph_proxy::api::graph::element::{Details, Element, GraphElement};
+    use graph_proxy::apis::{Details, Element, GraphElement};
     use graph_proxy::{create_demo_graph, SimplePartition};
     use ir_common::expr_parse::str_to_expr_pb;
     use ir_common::generated::algebra as pb;
