+++ research/query_service/ir/integrated/tests/graph_query_test.rs
@@ -21,7 +21,7 @@ mod common;
 mod test {
     use dyn_type::object;
     use dyn_type::Object;
-    use graph_proxy::api::graph::element::GraphElement;
+    use graph_proxy::apis::GraphElement;
     use ir_common::expr_parse::str_to_expr_pb;
     use ir_common::generated::algebra as pb;
     use ir_common::generated::common as common_pb;
