+++ research/query_service/ir/integrated/tests/pathxd_test.rs
@@ -19,8 +19,7 @@ mod common;
 
 #[cfg(test)]
 mod test {
-    use graph_proxy::api::graph::element::GraphElement;
-    use graph_proxy::api::graph::ID;
+    use graph_proxy::apis::{GraphElement, ID};
     use ir_common::generated::algebra as pb;
     use ir_common::generated::common as common_pb;
     use pegasus_client::builder::*;
