+++ research/query_service/ir/runtime/src/process/operator/group/group.rs
@@ -83,7 +83,7 @@ mod tests {
     use std::collections::{HashMap, HashSet};
 
     use dyn_type::Object;
-    use graph_proxy::api::graph::element::{DefaultDetails, DynDetails, Vertex};
+    use graph_proxy::apis::{DefaultDetails, DynDetails, Vertex};
     use ir_common::generated::algebra as pb;
     use ir_common::generated::common as common_pb;
     use ir_common::NameOrId;
