+++ research/query_service/ir/runtime/src/process/operator/sort/sort.rs
@@ -84,9 +84,7 @@ mod tests {
     use std::collections::HashMap;
 
     use dyn_type::Object;
-    use graph_proxy::api::graph::element::{
-        DefaultDetails, Details, DynDetails, Element, GraphElement, Vertex,
-    };
+    use graph_proxy::apis::{DefaultDetails, Details, DynDetails, Element, GraphElement, Vertex};
     use ir_common::generated::algebra as pb;
     use ir_common::generated::common as common_pb;
     use ir_common::NameOrId;
